// src/models/DonorRegistration.ts
export interface IDonorFileRef {
  fileId: string;
  url: string;
  name?: string;
  type?: string;
  fileName?: string;
  folder?: string;
  size?: number | string;
}